refactor(Problem7): extract methods to FriendRelations class
